feat: Add collapsible sections to the navStore
This commit adds the collapsible functionality to the Aside's "Subscripts" section. The active section's names are now stored in a newly added activeAsideCollapseName ref in the nav store. Additionally, the condition for showing the "Show More" button has been updated to avoid it being displayed when there are no items to load.