fix: Resolve Issue with Infinite Scroll and Safari's Popover
This commit fixes two issues. Firstly, it resolves an issue with infinite-scroll
in which the infinite-scroll-immediate attribute was being set to a string value
instead of a boolean. Secondly, it adds a "name" attribute to the Transition
element in the UserAvatar.vue component for Safari compatibility.
The name value is set to "nomean".