pref: add RouterLink to User Info in WatchMain component
This commit adds a RouterLink to the user profile picture and name in the
WatchMain.vue component. Clicking on either will direct the user to the
corresponding user's space page. The change is made by wrapping the ElImage
and span elements in separate RouterLink. Additionally, the user profile
picture now turns into a pointer cursor when hovered over.