feat: Add ownerId to VideoCard and VideosBoard components, Add pointer cursor to HeaderSearch
Refactor VideoCard and VideosBoard components to include the ownerId in
the props. The ownerId is obtained from the 'mid' property of the 'video.owner'
object. Additionally, the ownerId is used as a prop in RouterLink, which is
linked to the owner's space page.

Added the 'pointer' cursor style to the header-search el-input-group__append
class to improve the user experience when hovering over the search button.