feat: Fetch bilibili when mounting to get default cookie
Added a fetch request for /bilibili in App.vue on mounting to get the default
cookie from bilibili.com.

pref: Remove prop from ElEmpty component

In MyEmpty.vue, removed description prop from the ElEmpty component since it is
not being used.

feat: Add beforeEnter guard to protected pages

Added a beforeEnter guard to routes that render protected pages. If the user is
not logged in, a message is displayed prompting them to log in and the route is
not loaded.

feat: Add proxy configuration for bilibili

In vite.config.ts, added proxy configuration for /bilibili requests. Converted
the repeated configuration code to a function and added an optional parameter
to replace the path for specific cookies.